python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "asyncio>=3.4.3",
    "elevenlabs>=2.1.0",
    "fal-client>=0.7.0",
    "fastapi>=0.115.12",
    "google-genai>=1.18.0",
    "moviepy>=2.2.1",
    "python-dotenv>=1.1.0",
    "sievedata>=1.5.0",
    "supabase>=2.15.2",
    "uvicorn>=0.34.2",
]
